# docs/swagger.yaml
basePath: /
...
info:
  contact:
    url: https://github.com/0xbarchitect
  description: All-in-one SSO Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSO
  version: "1.0"
paths:
...
  /authorized/google:
    get:
      operationId: GoogleAuthorizedReq
      parameters:
      - description: state
        in: query
        name: state
        required: true
        type: string
      - description: redirect_url
        in: query
        name: redirect_url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: GoogleAuthorizedReq
  /authorized/google_handler:
    get:
      operationId: GoogleOauthHandler
      parameters:
      - description: state
        in: query
        name: state
        required: true
        type: string
      - description: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: GoogleOauthHandler
...
  /connect/:provider:
    get:
      operationId: ConnectIdentityProvider
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JwtHeader: []
      summary: ConnectIdentityProvider
  /connect/google_handler:
    get:
      operationId: GoogleConnectHandler
      parameters:
      - description: state
        in: formData
        name: state
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: GoogleConnectHandler
...
securityDefinitions:
  ClientBasicAuth:
    type: basic
  JwtHeader:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"